ASoC: codecs: cs5638: Try resuming multiple times to mitigate ESD
This chip is particularly vulnerable to ESD events knocking it out for a
little bit. This should of course be fixed using proper hardware design,
but we can add a bit of lenience to the driver by trying a few times to
resume instead of immediately failing on the first try.

Signed-off-by: John Watts <contact@jookia.org>