clk: sunxi-ng: Consistently allow children to set audio PLL clocks
Most chips have the clocking code set up so that changing an audio clock
for I2S, SPDIF or other audio peripherals can influence the parent
clocks. This allows a wider variety of divisors for audio clocking.

The code for the D1, A100, H6, H616 and F1C100S do not do this.
As far as I can tell this is an oversight, not a chip limitation.
Enable it on these chips.

Signed-off-by: John Watts <contact@jookia.org>