clk: sunxi-ng: Limit audio PLL clock multipliers on some chips
The D1, A100 and H616 all need their N clock value limited to values
specified in the datasheet. If invalid multipliers are used then
incorrect frequencies are silently used.

PLL_AUDIO1 on the D1 is a guess based on information in the datasheet.
I suspect the actual safe range is 7.5<=N<=145 and 12<=N, like
PLL_AUDIO0 but N limited to 145 (24 * 145 is around 3.5GHz) based on
its operating range.

Signed-off-by: John Watts <contact@jookia.org>