clk: sunxi-ng: Do not use pll-audio-4x as an I2S parent on the T113
The T113 (and I assume the D1) seems to have trouble when using the
pll-audio-4x clock as a parent with certain rates.
With 8 channel TDM at 32 bits per channel, asking for a rate of 215 kHz
requests a 55.04 MHz clock. The driver will set the registers correctly
but trying to use this clock gives an I/O error as the BCLK is not
output, instead a kind of flat line is emitted. The frame clock is still
correctly output, so this seems to be something to do with outputting
the BCLK in the I2S peripheral.

Here are the properties for that failure mode:

  i2s2 rate is 55024390
  i2s2 parent is pll-audio0-4x
  audio0 rate is 13756097
  audio0-4x rate is 55024390
  clk register dump:
    PLL_AUDIO0 0xF0285D00
    PLL_AUDIO0 PAT0 0x00000000
    PLL_AUDIO0 PAT1 0x00000000
    I2S2 0x01000000
  registers:
    sdm_enabled: 0
    p: 40
    n: 93
    input_div2: 0
    output_div2: 0
    wave_step: 0
    wave_bottom: 0
  VCO clk max: 2256.0 MHz
  final PLL clk: 55.02439024390244 to 55.02439024390244 MHz

Asking for rate 215.1 kHz outputs what looks like a high signal that is
a flat line with some bumps.  This no longer gives an error but instead
empty data from the codec.
Here are the properties for that failure mode:

  i2s2 rate is 55058823
  i2s2 parent is pll-audio0-4x
  audio0 rate is 13764705
  audio0-4x rate is 55058823
  clk register dump:
    PLL_AUDIO0 0xF8102600
    PLL_AUDIO0 PAT0 0x00000000
    PLL_AUDIO0 PAT1 0x00000000
    I2S2 0x81000000
  registers:
    sdm_enabled: 0
    p: 16
    n: 38
    input_div2: 0
    output_div2: 0
    wave_step: 0
    wave_bottom: 0
  VCO clk max: 936.0 MHz
  final PLL clk: 55.05882352941176 to 55.05882352941176 MHz

Nothing stands out to me about why this is happening, especially since
using pll-audio works just fine. I'm going to just disable this clock
parent for now and hope this is some unknown hardware errata.

Signed-off-by: John Watts <contact@jookia.org>